requirements.txt
Pillow
pywin32
requests
win32gui
